revert back axios instance interceptors

+++ src/utils/ApiInstance.js
@@ -6,10 +6,6 @@ import {getValue} from './sharedPreferences';
 
 const instance = axios.create({
   baseURL: '',
-  headers: {
-    Accept: 'application/json',
-    'Content-Type': 'application/json',
-  },
 });
 
 instance.interceptors.request.use(
@@ -54,35 +50,25 @@ instance.interceptors.response.use(
     console.log('====================================');
     console.log('errp==>', error);
     console.log('====================================');
-
     let {statusCode} = error?.response?.data;
-
-    switch (statusCode) {
-      case 401: {
-        if (global.is401Navigated === false) {
-          global.is401Navigated = true;
-          await AsyncStorage.clear();
-          ToastMessage({message: error?.response?.data?.message});
-          navigationRef?.current?.reset({
-            index: 0,
-            routes: [{name: 'OnBoarding'}],
-          });
-        }
-        break;
-      }
-
-      case 404:
-      case 500: {
+    if (statusCode === 401) {
+      if (global.is401Navigated === false) {
+        global.is401Navigated = true;
+        await AsyncStorage.clear();
         ToastMessage({message: error?.response?.data?.message});
         navigationRef?.current?.reset({
           index: 0,
-          routes: [{name: 'Main'}],
+          routes: [{name: 'OnBoarding'}],
         });
-        break;
       }
-
-      default:
-        Promise.reject(error.response.data);
+    } else if (statusCode === 404 || statusCode === 500) {
+      ToastMessage({message: error?.response?.data?.message});
+      navigationRef?.current?.reset({
+        index: 0,
+        routes: [{name: 'Main'}],
+      });
+    } else {
+      return Promise.reject(error.response.data);
     }
   },
 );
